Add: Extension.swift file && Move: HomeModel's extension file to Extentions file
