+++ Hellth/Hellth/Hellth/Home/Views/CalorieCircleView.swift
@@ -2,13 +2,12 @@ import SwiftUI
 
 struct CalorieCircleView: View {
     var totalCalories: Int = 1000
-    var consumedCalories: Int = 800
+    var consumedCalories: Int = 600
     var remainingCalories: Int {
         return totalCalories - consumedCalories
     }
     var body: some View {
         let progress = Double(consumedCalories) / Double(totalCalories)
-        let randomColors: [Color] = [ .red, .orange, .yellow, .green, .blue, .purple, .pink, .teal, .brown, .gray, .black, .white, .mint, .indigo]
         
         ZStack {
             Circle()
@@ -27,7 +26,7 @@ struct CalorieCircleView: View {
                 )
             Circle()
                 .trim(from: 0, to: CGFloat(progress))
-                .stroke(AngularGradient(gradient: Gradient(colors:randomColors.shuffled()), center: .center), lineWidth: 10)
+                .stroke(AngularGradient(gradient: Gradient(colors:Color.randomColors.shuffled()), center: .center), lineWidth: 10)
                 .rotationEffect(.degrees(-90))
                 .animation(.easeInOut(duration: 1.0), value: progress)
                 .shadow(color: .gray, radius: 3, x: 8, y: 8)
